requirements.txt
requests
newspaper3k
transformers
torch
lxml_html_clean